requirements.txt
# Web Interface Dependencies
streamlit==1.28.0
plotly==5.17.0
pandas==2.1.3
numpy==1.24.3

# Core NLP Dependencies (same as before)
spacy==3.7.2
transformers==4.36.0
torch==2.1.0
nltk==3.8.1
pyyaml==6.0.1

# For file uploads (optional)
python-docx==0.8.11